requirements.txt
streamlit==1.20.0
plotly==5.13.1
pandas==1.4.0
numpy==1.22.1
folium==0.14.0
streamlit_folium==0.11.1
Pillow==7.0.0
inflection==0.5.1
CurrencyConverter==0.17.5
plotly==5.13.1